+++ src/main/java/com/github/bartimaeusnek/bartworks/common/tileentities/multis/GT_TileEntity_THTR.java
@@ -248,11 +248,11 @@ public class GT_TileEntity_THTR extends GT_MetaTileEntity_EnhancedMultiBlockBase
         if (!(this.HeliumSupply >= GT_TileEntity_THTR.HELIUM_NEEDED && this.fuelsupply >= mincapacity)) return false;
 
         double eff = Math.min(
-                Math.pow((double) (this.fuelsupply - mincapacity) / ((maxcapacity - mincapacity) / 10D), 2D) + 1,
-                100D) / 100D - ((double) (this.getIdealStatus() - this.getRepairStatus()) / 10D);
+                Math.pow((this.fuelsupply - mincapacity) / ((maxcapacity - mincapacity) / 10D), 2D) + 1,
+                100D) / 100D - ((this.getIdealStatus() - this.getRepairStatus()) / 10D);
         if (eff <= 0D) return false;
 
-        int toReduce = MathUtils.floorInt((double) this.fuelsupply * 0.005D * eff);
+        int toReduce = MathUtils.floorInt(this.fuelsupply * 0.005D * eff);
 
         final int originalToReduce = toReduce;
         int burnedballs = toReduce / 64;
