+++ src/main/java/com/github/bartimaeusnek/crossmod/GTpp/loader/RadioHatchCompat.java
@@ -226,7 +226,7 @@ public class RadioHatchCompat {
             try {
                 List<?> pureElements = RadioHatchCompat.GTPPRadAdapter.getElemets(m);
                 for (Object materialObj : pureElements) if (RadioHatchCompat.isRadioactive.getBoolean(materialObj))
-                    ret += ((int) RadioHatchCompat.radlevel.getByte(m) + RadioHatchCompat.GTPPRadAdapter
+                    ret += (RadioHatchCompat.radlevel.getByte(m) + RadioHatchCompat.GTPPRadAdapter
                             .clampToZero(RadioHatchCompat.protons.getLong(materialObj)));
                 else ret += RadioHatchCompat.radlevel.getByte(m);
             } catch (IllegalAccessException e) {
